Python/document_intelligence/requirements.txt
Flask==2.1.3
python-docx==0.8.11
PyMuPDF==1.18.19
Pillow==9.0.0
pytesseract==0.3.8
pip install spacy
python -m spacy download en_core_web_sm